repo reorg

+++ StaticGrad/include/utils.hpp
@@ -2,14 +2,12 @@
 #define UTILS_HPP
 
 #include "./classes.hpp"
-#include <random>
-#include <fstream>
 
 void crossentropy_forward(Node* out, Node* in, u_int16_t* targets);
 
 void crossentropy_softmax_backward(Node* out, Node* in, u_int16_t* targets, float temperature);
 
-int sample_token(float* probabilities, int length, bool random = false);
+u_int16_t sample_token(float* probabilities, size_t length, bool random = false);
 
 void load_weights(float* dest, const std::string& fname, int expected_bytes = -1);
